worker-service/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic>=2.0
google-genai>=1.0.0
requests>=2.31.0
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-instrumentation-requests==0.46b0
opentelemetry-instrumentation-grpc==0.46b0
kafka-python==2.0.2
psycopg2-binary==2.9.9
pytest==8.3.3
pytest-asyncio==0.23.0
httpx==0.27.2
PyJWT==2.8.0
pyyaml==6.0.1
